tree queries

+++ templates/src/dset.rs
@@ -2,24 +2,24 @@ mod collections {
     use std::cell::Cell;
 
     pub struct DisjointSet {
-        parent: Vec<Cell<usize>>,
+        parent: Vec<Cell<u32>>,
         size: Vec<u32>,
     }
 
     impl DisjointSet {
         pub fn new(n: usize) -> Self {
             Self {
-                parent: (0..n).map(|i| Cell::new(i)).collect(),
+                parent: (0..n as u32).map(|i| Cell::new(i)).collect(),
                 size: vec![1; n],
             }
         }
 
         pub fn find_root(&self, u: usize) -> usize {
-            if u == self.parent[u].get() {
+            if u == self.parent[u].get() as usize {
                 u
             } else {
-                self.parent[u].set(self.find_root(self.parent[u].get()));
-                self.parent[u].get()
+                self.parent[u].set(self.find_root(self.parent[u].get() as usize) as u32);
+                self.parent[u].get() as usize
             }
         }
 
@@ -37,7 +37,7 @@ mod collections {
             if self.size[u] > self.size[v] {
                 std::mem::swap(&mut u, &mut v);
             }
-            self.parent[v].set(u);
+            self.parent[v].set(u as u32);
             self.size[u] += self.size[v];
             true
         }
